调整函数
